video_deepfake/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
timm>=0.9.0
facenet-pytorch>=2.5.2
numpy>=1.24.0
pillow>=10.0.0
